refactor(admin): remove the retired ISF client layer
ISF is gone (docs/exec-plans/admin-bkn-safe-migration.md): the `admin`
group moved onto bkn-safe's `/api/safe/v1/admin/*`, and `src/api/safe.ts`
has backed every command since. The old clients for the four ISF base
paths — /api/user-management/v1, /api/authorization/v1,
/isfweb/api/ShareMgnt, /api/eacp/v1 — stayed behind, unreachable: of the
21 runtime exports in api/admin.ts, 20 had no caller anywhere in src, and
`resources/admin.ts` imports only types from it. All 9 types are still
live, so api/admin.ts keeps them and loses everything else.

Also removed:

- eacp-crypto.ts. Its only callers were the two dead password functions,
  so the whole file goes, and with it the 1024-bit RSA / PKCS#1 v1.5
  surface — both below current guidance, and the reason a private key was
  embedded here in the first place (28f359e cut the key; this cuts the
  need for it).
- org-tree.ts's `buildOrgTree`, which nested ISF `parent_deps[]` chains.
  bkn-safe returns `parent_id` and `buildDepartmentTree` in api/safe.ts
  builds the tree; `renderOrgTree` stays and is fed by that.
- test/unit/admin.test.ts, whose 12 cases all exercised removed
  functions against ISF routes.

`changePasswordSafe` was the one live export in api/admin.ts and never
belonged there — it targets bkn-safe, not ISF. Moved to api/safe.ts
alongside the rest.

Verified against a mock server, driving the built CLI: `auth
change-password` still POSTs /api/safe/v1/auth/change-password with the
same body and returns ok; `admin org tree` renders through
buildDepartmentTree + renderOrgTree; `user list`, `user reset-password`
(plaintext, no RSA), and `role list` hit their /api/safe/v1/admin/*
routes; `audit list` still gives the notOnSafe error. No ISF path or RSA
call remains in src or dist.
